test(LottoTest): 테스트 클래스를 도메인 패키지로 이동
