+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,0 +1,40 @@
+package lotto.domain;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import java.util.List;
+import lotto.domain.Lotto;
+import lotto.domain.Number;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class LottoTest {
+    @DisplayName("로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.")
+    @Test
+    void createLottoByOverSize() {
+        assertThatThrownBy(() -> Lotto.from(List.of(
+                Number.valueOf(1),
+                Number.valueOf(2),
+                Number.valueOf(3),
+                Number.valueOf(4),
+                Number.valueOf(5),
+                Number.valueOf(6),
+                Number.valueOf(7)
+        )))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("로또 번호에 중복된 숫자가 있으면 예외가 발생한다.")
+    @Test
+    void createLottoByDuplicatedNumber() {
+        assertThatThrownBy(() -> Lotto.from(List.of(
+                Number.valueOf(1),
+                Number.valueOf(2),
+                Number.valueOf(3),
+                Number.valueOf(4),
+                Number.valueOf(5),
+                Number.valueOf(5)
+        )))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+}
